Udacity Sensor Fusion: added function that helps to compute #points in segmentations
The segmentation functions take arguments like #min points and #max points
to determine the smallest and biggest cluster. This depends also on the
used Lidar (in particular its resolution). With the helper function, this
can be simplified using a density function that is used to scale the
forehand mentioned values.