Udacity Sensor Fusion: removed the shared() part from the OMP statement
Older compiler version seem to have issues to handle the context, because
it does not allow to use variables from structures in shared statement.
The default usage of those variables is shared anyway, and therefore, the
removal has not effect (beside it allows to compile the code & run it
parallel on older compilers).